% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_days_since}
\alias{add_days_since}
\title{Helper function}
\usage{
add_days_since(data, start_cutoff)
}
\arguments{
\item{data}{data frame with `region`, `Date` and `total` columns}
}
\value{
data fram with extra column `d` with days since `total` reached `start_cutoff`
}
\description{
Helper function
}
