% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_rolling_growth_rates}
\alias{compute_rolling_growth_rates}
\title{compute rolling growth rates}
\usage{
compute_rolling_growth_rates(data, window_width)
}
\arguments{
\item{data}{data frame with columns `region`, `total` and `d` to compute rolling growth rates by region
using the total variable at time steps d}

\item{window_width}{of window for rolling fit}
}
\value{
the original data frame with added columns `slope` (growth rate), `slope_e` (standard error of growth rate fit),
`low` and `high` (the 95% confidencit interval of the fit)
}
\description{
compute rolling growth rates
}
