% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_alberta_case_data}
\alias{get_alberta_case_data}
\title{import and recode ontario case data from Alberta Open Data. Tends to have a day lag}
\usage{
get_alberta_case_data()
}
\value{
a wide format data frame with one row per case with Health Region, gender, age group,
status, case type
}
\description{
import and recode ontario case data from Alberta Open Data. Tends to have a day lag
}
