% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_bc_covid_speak_survey_pop_group_data}
\alias{get_bc_covid_speak_survey_pop_group_data}
\title{get BC Covid SPEAK population group survey data}
\usage{
get_bc_covid_speak_survey_pop_group_data()
}
\value{
a dataframe with data
}
\description{
get BC Covid SPEAK population group survey data
}
