% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_british_columbia_case_data}
\alias{get_british_columbia_case_data}
\title{import and recode case data from British Columbia CDC. Tends to have a day lag}
\usage{
get_british_columbia_case_data()
}
\value{
a wide format data frame with one row per case with Health Authority, gender, age group
and report date
}
\description{
import and recode case data from British Columbia CDC. Tends to have a day lag
}
