% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_british_columbia_hr_case_data}
\alias{get_british_columbia_hr_case_data}
\title{import and recode Health Region (Health Service Delivery Area) case data from British Columbia CDC. Tends to have a day lag}
\usage{
get_british_columbia_hr_case_data()
}
\value{
a wide format data frame with one row per Date, Health Authority, Health Region,
with Cases and Cases Smoothed
}
\description{
import and recode Health Region (Health Service Delivery Area) case data from British Columbia CDC. Tends to have a day lag
}
