% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_british_columbia_test_data}
\alias{get_british_columbia_test_data}
\title{import and recode test data from British Columbia CDC. Tends to have a day lag}
\usage{
get_british_columbia_test_data()
}
\value{
a wide format data frame with one row per date and Health Authority, with additional columns
New Tests, Total Tests, Positivity and Trun Around
}
\description{
import and recode test data from British Columbia CDC. Tends to have a day lag
}
