% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_can_covid_tracker_data}
\alias{get_can_covid_tracker_data}
\title{import and recode case data from caanada covid tracker}
\usage{
get_can_covid_tracker_data(province)
}
\arguments{
\item{province}{two letter provincial abbreviation code}
}
\value{
a wide format data frame
}
\description{
import and recode case data from caanada covid tracker
}
