% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_combined_provincial_data}
\alias{get_canada_combined_provincial_data}
\title{combined data from UofS, Canada.ca and Canada Covid-19 working group.}
\usage{
get_canada_combined_provincial_data()
}
\value{
dataframe with columns `shortProvince`, `Province`, `Date`,
`Confirmed`, `Deaths`, `Cases`, `Recovered`, `Active`
}
\description{
combined data from UofS, Canada.ca and Canada Covid-19 working group.
}
