% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_cases}
\alias{get_canada_covid_working_group_cases}
\title{Get confirmed case data from the covid-19 data working group}
\usage{
get_canada_covid_working_group_cases()
}
\value{
tibble with case data
}
\description{
Get confirmed case data from the covid-19 data working group
}
