% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_data}
\alias{get_canada_covid_working_group_data}
\title{data from the covid-19 data working group (https://github.com/ishaberry/Covid19Canada)}
\usage{
get_canada_covid_working_group_data(use_csv = FALSE)
}
\arguments{
\item{use_csv}{use convenience csv download instead of cases}
}
\value{
dataframe with columns `province`, `health_region`, `Date`, `type`, `count`
`type` is "Cases", "Deaths", "Recovered". All are new daily cases
and Recovered is cumulative recovered cases.
}
\description{
data from the covid-19 data working group (https://github.com/ishaberry/Covid19Canada)
}
