% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_deaths}
\alias{get_canada_covid_working_group_deaths}
\title{Get deaths data from the covid-19 data working group}
\usage{
get_canada_covid_working_group_deaths()
}
\value{
tibble with deaths
}
\description{
Get deaths data from the covid-19 data working group
}
