% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_health_region_data}
\alias{get_canada_covid_working_group_health_region_data}
\title{data from the covid-19 data working group}
\usage{
get_canada_covid_working_group_health_region_data(
  start_cutoff,
  cache_key,
  refresh = FALSE
)
}
\value{
dataframe with columns `Health Region`, `Province`, `shortProvince`, `Date`,
`Confirmed`, `Deaths`, where all counts are cumulative,
}
\description{
data from the covid-19 data working group
}
