% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_provincial_data}
\alias{get_canada_covid_working_group_provincial_data}
\title{data from the covid-19 data working group (https://github.com/ishaberry/Covid19Canada)}
\usage{
get_canada_covid_working_group_provincial_data()
}
\value{
dataframe with columns `Province`, `shortProvince`, `Date`,
`Confirmed`, `Deaths`, `Recovered`, `Active`, where all counts are cumulative,
}
\description{
data from the covid-19 data working group (https://github.com/ishaberry/Covid19Canada)
}
