% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_recovered}
\alias{get_canada_covid_working_group_recovered}
\title{Get recovered case data from the covid-19 data working group}
\usage{
get_canada_covid_working_group_recovered()
}
\value{
tibble with recovered case data
}
\description{
Get recovered case data from the covid-19 data working group
}
