% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_timeseries}
\alias{get_canada_covid_working_group_timeseries}
\title{Get Open Canada Working Group timeseries data via API}
\usage{
get_canada_covid_working_group_timeseries(type = "cases", location = "prov")
}
\arguments{
\item{type}{type of data, one of cases, mortality, recovered, testing, active, avaccine, dvaccine or cvaccine}

\item{location}{of data, either level of geography canada, prov, hr, or a specific region code, two-letter province code or a health region code}
}
\value{
a data frame with the data
}
\description{
Get Open Canada Working Group timeseries data via API
}
