% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_UofS_case_data}
\alias{get_canada_UofS_case_data}
\title{data from UofS}
\usage{
get_canada_UofS_case_data()
}
\value{
dataframe with columns `Province`, `shortProvince`, `Health Region`, `Date`, `Age`,
`Travel history`, `Confirmed state`
}
\description{
data from UofS
}
