% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_cansim_case_data}
\alias{get_cansim_case_data}
\title{import and recode canadian case data from StatCan table 13-10-0767}
\usage{
get_cansim_case_data()
}
\value{
a wide format data frame with one row per case with gender, age group,
hospitilization and ICU status, transmission pathway, mortality, and onset of symptoms data
Data is limited to case information forwarded by provinces to Canada Health
}
\description{
import and recode canadian case data from StatCan table 13-10-0767
}
