% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/international_import.R
\name{get_ecdc_data}
\alias{get_ecdc_data}
\title{Import ecdc case data}
\usage{
get_ecdc_data()
}
\value{
dataframe with columns `Country`, `shortCountry`, `countryCode`, `Population 2018`, `Date`,
`Confirmed`, `Deaths`, `Cases`, where `Cases` is daily confirmed case count increase and
`Confirmed` and `Deaths` are cumulative
}
\description{
Import ecdc case data
}
