% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_health_region_census_2016_data}
\alias{get_health_region_census_2016_data}
\title{get Health Region level census data}
\usage{
get_health_region_census_2016_data(refresh = FALSE)
}
\arguments{
\item{refresh}{set to `TRUE` to refresh locally cached data}
}
\value{
a long form dataframe with census data for health regions
The health region geography used for this extract is different from the 2018 health regions
as it misses at least one health region amalgamation in 2018 (and the one in 2020) in Ontario.
}
\description{
get Health Region level census data
}
