% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_health_region_geographies_2018}
\alias{get_health_region_geographies_2018}
\title{get Health Region geographies}
\usage{
get_health_region_geographies_2018(refresh = FALSE)
}
\arguments{
\item{refresh}{data is cached for the duration of the session, set to `TRUE` to refresh the cache}
}
\value{
a simple feature collection with 2018 Health Region data
This misses at least one 2020 change to Ontario Health regoins
}
\description{
get Health Region geographies
}
