% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_manitoba_daily_case_data}
\alias{get_manitoba_daily_case_data}
\title{import and recode case data from Saskatchewan.}
\usage{
get_manitoba_daily_case_data(date = NULL)
}
\arguments{
\item{date}{for which to download the data. Defauly is NULL and downloads the newest available}
}
\value{
a wide format data frame
}
\description{
import and recode case data from Saskatchewan.
}
