% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_ontario_case_data}
\alias{get_ontario_case_data}
\title{import and recode ontario case data from Ontario Open Data. Tends to have a day lag}
\usage{
get_ontario_case_data()
}
\value{
a wide format data frame with one row per case with Health Region, gender, age group,
transmission pathway, status, and onset of symptoms data
}
\description{
import and recode ontario case data from Ontario Open Data. Tends to have a day lag
}
