% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_open_table_data}
\alias{get_open_table_data}
\title{import open table data from \link{https://www.opentable.com/state-of-industry}}
\usage{
get_open_table_data(type = c("fullbook", "reopening", "occupancy"))
}
\arguments{
\item{type, }{type of data, one of "fullbook", "reopening", or "occupancy"}
}
\value{
open table booking, opening or occupancy data
}
\description{
import open table data from \link{https://www.opentable.com/state-of-industry}
}
