% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_quebec_case_data}
\alias{get_quebec_case_data}
\title{import and recode case data from Quevec}
\usage{
get_quebec_case_data()
}
\value{
a wide format data frame
}
\description{
import and recode case data from Quevec
}
