% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_toronto_neighbourhood_cases}
\alias{get_toronto_neighbourhood_cases}
\title{get Toronto case data}
\usage{
get_toronto_neighbourhood_cases()
}
\value{
a data frame with Toronto neighbourhood level covid-19 counts
}
\description{
get Toronto case data
}
