% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_toronto_neighbourhood_cases_old}
\alias{get_toronto_neighbourhood_cases_old}
\title{get Toronto case data old}
\usage{
get_toronto_neighbourhood_cases_old()
}
\value{
a data frame with Toronto neighbourhood level covid-19 counts
}
\description{
get Toronto case data old
}
