% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_toronto_neighbourhood_census_data}
\alias{get_toronto_neighbourhood_census_data}
\title{get Toronto neihgbourhood census data 2016}
\usage{
get_toronto_neighbourhood_census_data(refresh = FALSE)
}
\arguments{
\item{refresh}{data is cached for the duration of the session, set to `TRUE` to refresh the cache}
}
\value{
a data frame with toronto neighbourhood census data
}
\description{
get Toronto neihgbourhood census data 2016
}
