% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_toronto_neighbourhood_geos}
\alias{get_toronto_neighbourhood_geos}
\title{get Toronto neihgbourhood geographies}
\usage{
get_toronto_neighbourhood_geos(refresh = FALSE)
}
\arguments{
\item{refresh}{data is cached for the duration of the session, set to `TRUE` to refresh the cache}
}
\value{
a simple feature collection with Toronto neighbrouhoods
}
\description{
get Toronto neihgbourhood geographies
}
