% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_graphs.R
\name{growth_rate_graph}
\alias{growth_rate_graph}
\title{Generate grawth rate graph}
\usage{
growth_rate_graph(data, window_width)
}
\arguments{
\item{data}{data fram with colum `region` for each group, `Date` with dates and `total`, for fitting the growth rates}
}
\value{
a `ggplot` object
}
\description{
Generate grawth rate graph
}
