% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{replace_ON_health_region_geocodes}
\alias{replace_ON_health_region_geocodes}
\title{aggregate health regions to health autorities in ON}
\usage{
replace_ON_health_region_geocodes(data)
}
\arguments{
\item{data}{dataframe with fields `GeoUID` = HR_UID and `Name`}
}
\value{
a dataframe with GeoUID and Name adjusted to current names
}
\description{
aggregate health regions to health autorities in ON
}
