% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simpleCache}
\alias{simpleCache}
\title{Simple key-value cache function accepting closures}
\usage{
simpleCache(object, key, path = getOption("custom_data_path"), refresh = FALSE)
}
\arguments{
\item{object}{closure with return expression to be cached}

\item{key}{cache key}

\item{path}{path to cache the data}

\item{refresh}{bool option to force refresh of cache, default FALSE}
}
\description{
Simple key-value cache function accepting closures
}
