% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_data}
\alias{get_canada_covid_working_group_data}
\title{data from the covid-19 data working group}
\usage{
get_canada_covid_working_group_data()
}
\value{
dataframe with columns `province`, `health_region`, `Date`, `type`, `count`
`type` is "Cases", "Deaths", "Recovered". All are new daily cases
and Recovered is cumulative recovered cases.
}
\description{
data from the covid-19 data working group
}
