% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_covid_working_group_tests}
\alias{get_canada_covid_working_group_tests}
\title{Get data on testing in Canada}
\usage{
get_canada_covid_working_group_tests()
}
\value{
tibble with number of tests by province and data
}
\description{
Get data on testing in Canada
}
