% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada_covid_data_import.R
\name{get_canada_UofS_health_region_data}
\alias{get_canada_UofS_health_region_data}
\title{data from (https://covid19tracker.ca/)}
\usage{
get_canada_UofS_health_region_data()
}
\value{
dataframe with columns `PR_UID`, `prname`, `prnameFR`, `shortProvince`, `Date`,
`Confirmed`, `Probable`, `Total`, `Deaths`, `Cases`, `Tested` where `Cases` is new daily cases and
all other counts are cumulative, `Total` is `Confirmed` pluis `Probable` (postive tests which have not been confirmed at an official national or provincial lab)
}
\description{
data from (https://covid19tracker.ca/)
}
