% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/international_import.R
\name{get_country_timeline_ecdc_jhs_data}
\alias{get_country_timeline_ecdc_jhs_data}
\title{Import country level data, merge ecdc and jhs timelines}
\usage{
get_country_timeline_ecdc_jhs_data()
}
\value{
dataframe with columns `Country`, `Date`,
`Deaths`, `Recovered`, `Active`, `Cases`, Cases is daily new cases, others are cumulative
}
\description{
Import country level data, merge ecdc and jhs timelines
}
